# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: 給与明細・バックオフィス効率化システム API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /payroll:
    get:
      operationId: listPayrolls
      tags: [payroll]
      summary: 給与一覧
      parameters:
        - name: year
          in: query
          required: true
          schema:
            type: integer
        - name: month
          in: query
          required: true
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, confirmed]
      responses:
        "200":
          description: 給与一覧
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payroll"
  /payroll/calculate:
    post:
      operationId: calculatePayroll
      tags: [payroll]
      summary: 給与計算実行
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CalculatePayrollBody"
      responses:
        "200":
          description: 計算結果
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payroll"
  /payroll/{id}:
    get:
      operationId: getPayroll
      tags: [payroll]
      summary: 給与明細詳細
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: 給与明細
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payroll"
    put:
      operationId: updatePayroll
      tags: [payroll]
      summary: 給与明細更新
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePayrollBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payroll"
  /payroll/{id}/confirm:
    post:
      operationId: confirmPayroll
      tags: [payroll]
      summary: 給与確定
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payroll"
  /payroll/{id}/csv:
    get:
      operationId: exportPayrollCsv
      tags: [payroll]
      summary: 給与明細CSV出力
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: CSV data
          content:
            text/csv:
              schema:
                type: string

  # ── 振替伝票 ──────────────────────────────────────────────────────────────
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
  schemas:
...
    Payroll:
      type: object
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        employeeName:
          type: string
        employeeCode:
          type: string
        year:
          type: integer
        month:
          type: integer
        status:
          type: string
          enum: [draft, confirmed]
        baseSalary:
          type: number
          description: 基本給
        overtimePay:
          type: number
          description: 時間外手当
        lateNightPay:
          type: number
          description: 深夜手当
        holidayPay:
          type: number
          description: 休日手当
        commissionPay:
          type: number
          description: 歩合給
        transportationAllowance:
          type: number
          description: 通勤手当
        safetyDrivingAllowance:
          type: number
          description: 無事故手当
        longDistanceAllowance:
          type: number
          description: 長距離手当
        positionAllowance:
          type: number
          description: 役職手当
        familyAllowance:
          type: number
          description: 家族手当
        earlyOvertimeAllowance:
          type: number
          description: 早出残業手当
        absenceDeduction:
          type: number
          description: 欠勤控除
        grossSalary:
          type: number
          description: 支給合計
        socialInsurance:
          type: number
          description: 社会保険料
        employmentInsurance:
          type: number
          description: 雇用保険料
        incomeTax:
          type: number
          description: 源泉所得税
        residentTax:
          type: number
          description: 住民税
        totalDeductions:
          type: number
          description: 控除合計
        netSalary:
          type: number
          description: 差引支給額
        overtimeHours:
          type: number
        lateNightHours:
          type: number
        holidayWorkDays:
          type: number
        workDays:
          type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - employeeId
        - year
        - month
        - status
        - baseSalary
        - overtimePay
        - lateNightPay
        - holidayPay
        - commissionPay
        - transportationAllowance
        - safetyDrivingAllowance
        - longDistanceAllowance
        - positionAllowance
        - familyAllowance
        - earlyOvertimeAllowance
        - absenceDeduction
        - grossSalary
        - socialInsurance
        - employmentInsurance
        - incomeTax
        - residentTax
        - totalDeductions
        - netSalary
        - overtimeHours
        - lateNightHours
        - holidayWorkDays
        - workDays
        - createdAt
        - updatedAt
    UpdatePayrollBody:
      type: object
      properties:
        safetyDrivingAllowance:
          type: number
        longDistanceAllowance:
          type: number
        positionAllowance:
          type: number
        notes:
          type: string
    CalculatePayrollBody:
      type: object
      properties:
        employeeId:
          type: integer
        year:
          type: integer
        month:
          type: integer
      required:
        - employeeId
        - year
        - month
...